Split classes related to animations
