+++ lib/src/animations/spinner_curve.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class SpinnerCurve extends Curve {
+  const SpinnerCurve();
+
+  @override
+  double transform(double tr) => (tr <= 0.5) ? 1.9 * tr : 1.85 * (1 - tr);
+}
